PRIME | move config to conf.py
